% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example_survey.R
\docType{data}
\name{example_survey}
\alias{example_survey}
\title{Example Survey Dataset}
\format{
A data frame with 10 rows and 5 variables:
\describe{
  \item{Age}{Numeric age of respondent}
  \item{Gender}{Gender of respondent (Male/Female)}
  \item{Education}{Education level (High School/Bachelor/Graduate)}
  \item{Income}{Numeric income value}
  \item{Weight}{Survey weight}
}
}
\source{
Simulated data for demonstration purposes
}
\usage{
example_survey
}
\description{
A small example dataset used to demonstrate SurveyStat functions.
}
\keyword{datasets}
